requirements.txt
streamlit
streamlit-option-menu
sqlalchemy
bcrypt
python-dotenv
mercadopago
fastapi
uvicorn
pandas
matplotlib
